differentiator: visitor-simplify: implement expansion of "(a+b)*const"
This is done only when a suitable muldiv (i. e. constant * addsub) is a child
of another addsub.

Also, the addsub decomposition and folding has been slightly reworked. Analogous
rework will follow for muldiv decomposition, which will allow to implement
expansion of "(a*b)^const" in a similar way (removing the more hackish
implementation from the Node::Power simplifier).